date	2005.03.02.06.02.00;	author endre;	state Exp;
git-svn-id: https://svn.isis.vanderbilt.edu/MoBIES/UDM/trunk@2163 c0adbae9-c137-0410-b742-ec4b2d2ceded

+++ src/UdmOclPat/GOCL/src/OCLObjectExBasic.cpp
@@ -293,7 +293,7 @@ namespace OclBasic
 
 	void Integer::SetValue( long value )
 	{
-		SetValue( (double) value );
+		SetValue(  value );
 	}
 
 	Any* Integer::Clone() const
